Fixed bug in positioning displays
Happens when getDisplays() returns displays out of spatial order, and
there are 4 or more displays. Eg. when the 2nd display spatially is
returned by getDisplays() in the 4th position, the X and Y indices of
the 3rd screen returned by getDisplays will not be updated.

Cred goes to Gregg Van Hove (https://github.com/gvanhove) for spotting
it.